check.

+++ tests/python-gpu/test_gpu_with_sklearn.py
@@ -80,7 +80,9 @@ def test_categorical():
     from sklearn.datasets import load_svmlight_file
 
     data_dir = tm.data_dir(__file__)
-    X, y = load_svmlight_file(os.path.join(data_dir, "agaricus.txt.train"))
+    X, y = load_svmlight_file(
+        os.path.join(data_dir, "agaricus.txt.train"), dtype=np.float32
+    )
     clf = xgb.XGBClassifier(
         tree_method="gpu_hist",
         enable_categorical=True,
